% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plearn-package.R
\docType{package}
\name{plearn-package}
\alias{plearn}
\alias{plearn-package}
\title{plearn: Automate PrairieLearn Operations}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (YOUR-ORCID-ID)

}
\keyword{internal}
