% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-fields.R
\name{use_input_integer}
\alias{use_input_integer}
\title{Create an Integer Input}
\usage{
use_input_integer(name, title = NULL, topic = NULL, tags = NULL)
}
\arguments{
\item{name}{Instructor-facing name of the question. This will be appended
with the relevant question type.}

\item{title}{Student-facing name for the question.}

\item{topic}{Category to put the question under.}

\item{tags}{Subtags to identify material tested or technologies used.}
}
\description{
Establishes an integer input field that allows for the evaluation of
only whole numbers.
}
