% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-fields.R
\name{use_input_symbolic}
\alias{use_input_symbolic}
\title{Create a Symbolic Input}
\usage{
use_input_symbolic(name, title = NULL, topic = NULL, tags = NULL)
}
\arguments{
\item{name}{Instructor-facing name of the question. This will be appended
with the relevant question type.}

\item{title}{Student-facing name for the question.}

\item{topic}{Category to put the question under.}

\item{tags}{Subtags to identify material tested or technologies used.}
}
\description{
Establishes symbolic input field that allows for the evaluation of
a mathematical expression.
}
