% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-choice.R
\name{use_mc_fixed}
\alias{use_mc_fixed}
\title{Create a Multiple Choice Question with a Fixed Number of Responses}
\usage{
use_mc_fixed(name, title = NULL, topic = NULL, tags = NULL)
}
\arguments{
\item{name}{Instructor-facing name of the question. This will be appended
with the relevant question type.}

\item{title}{Student-facing name for the question.}

\item{topic}{Category to put the question under.}

\item{tags}{Subtags to identify material tested or technologies used.}
}
\description{
Establishes a multiple choice question with a fixed amount of responses.
}
