% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-selection.R
\name{use_ms_fixed}
\alias{use_ms_fixed}
\title{Create a Multiple Selection Question with a Fixed Number of Responses}
\usage{
use_ms_fixed(name, title = NULL, topic = NULL, tags = NULL)
}
\arguments{
\item{name}{Instructor-facing name of the question. This will be appended
with the relevant question type.}

\item{title}{Student-facing name for the question.}

\item{topic}{Category to put the question under.}

\item{tags}{Subtags to identify material tested or technologies used.}
}
\description{
Establishes a multiple selection question that allows for one or more
responses to be selected.
}
