% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plearn-package.R
\docType{package}
\name{plearn-package}
\alias{plearn}
\alias{plearn-package}
\title{plearn: Automate PrairieLearn Operations}
\description{
Provides question templates in a convenient form for deploying
    to PrairieLearn
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458)

}
\keyword{internal}
